Cable_paradas/requirements.txt
asgiref==3.8.1
contourpy==1.2.1
cycler==0.12.1
Django==5.0.6
et-xmlfile==1.1.0
fonttools==4.53.0
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==2.0.0
openpyxl==3.1.5
packaging==24.1
pillow==10.3.0
plotly==5.22.0
psycopg2-binary==2.9.9
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.16.0
sqlparse==0.5.0
tenacity==8.4.1
